Hot fixes in Saugmecon and Gavanelli.
Saugmecon:
Avoid calculating the objectives nadir values, it is faster to use their lower bound.
In the problem RCPSP with makespan and weighted earliness-tardiness (wet) is very
hard to find the nadir of the wet objective by optimizing the inverse wet.

Gavanelli:
We shoudl raise a issue in the choco repository, if there are two solutions with the same
values the one that is in the front is replaced by the other one, but this is not usefull as
the code is prepared to store just one solution per Pareto point.