Hot fix the timeout for Saugmecon, GIA, Disjunctive. Exhaustive condition
The timeout was decreased in the wrong way, leading to stopping the execution
before the real timeout was reached.

Implement a way to know if the algorithms have found the whole front. Before
it was determined for the solver timeout but on some occasions exhaustive was
wrongly set to true.

In Saugmecon the search for previous solutions is now implemented more simply. The solutions with
the efArray are inserted in a list. To know if a previous efArray produce solutions
that can satisfy the current efArray we have to search in the list, time O(n)