Fix error while adding points obtained during individual optimization
If the exploration is not exhaustive we check if the points found when the objectives were individually optimized are not dominated and not in the front.
If that is true we add those points to the approximation front. This was not well implemented, now it is.

Fix comparison against ideal to discard infeasible regions